+++ mycli/completion_refresher.py
@@ -3,7 +3,7 @@ from .packages.special.main import COMMANDS
 from collections import OrderedDict
 
 from .sqlcompleter import SQLCompleter
-from .sqlexecute import SQLExecute
+from .sqlexecute import SQLExecute, ServerSpecies
 
 class CompletionRefresher(object):
 
@@ -113,6 +113,8 @@ def refresh_users(completer, executor):
 @refresher('functions')
 def refresh_functions(completer, executor):
     completer.extend_functions(executor.functions())
+    if executor.server_info.species == ServerSpecies.TiDB:
+        completer.extend_functions(completer.tidb_functions, builtin=True)
 
 @refresher('special_commands')
 def refresh_special(completer, executor):
@@ -121,3 +123,8 @@ def refresh_special(completer, executor):
 @refresher('show_commands')
 def refresh_show_commands(completer, executor):
     completer.extend_show_items(executor.show_candidates())
+
+@refresher('keywords')
+def refresh_keywords(completer, executor):
+    if executor.server_info.species == ServerSpecies.TiDB:
+        completer.extend_keywords(completer.tidb_keywords, replace=True)