+++ mycli/main.py
@@ -24,6 +24,7 @@ from cli_helpers.tabular_output import preprocessors
 from cli_helpers.utils import strip_ansi
 import click
 import sqlparse
+import sqlglot
 from mycli.packages.parseutils import is_dropping_database, is_destructive
 from prompt_toolkit.completion import DynamicCompleter
 from prompt_toolkit.enums import DEFAULT_BUFFER, EditingMode
@@ -92,6 +93,7 @@ SUPPORT_INFO = (
 class MyCli(object):
 
     default_prompt = '\\t \\u@\\h:\\d> '
+    default_prompt_splitln = '\\u@\\h\\n(\\t):\\d>'
     max_len_prompt = 45
     defaults_suffix = None
 
@@ -123,6 +125,7 @@ class MyCli(object):
         self.logfile = logfile
         self.defaults_suffix = defaults_suffix
         self.login_path = login_path
+        self.toolbar_error_message = None
 
         # self.cnf_files is a class variable that stores the list of mysql
         # config files to read in at launch.
@@ -425,6 +428,7 @@ class MyCli(object):
             port = 3306
             if not host or host == 'localhost':
                 socket = (
+                    socket or
                     cnf['socket'] or
                     cnf['default_socket'] or
                     guess_socket_location()
@@ -582,6 +586,34 @@ class MyCli(object):
             return True
         return False
 
+    def handle_prettify_binding(self, text):
+        try:
+            statements = sqlglot.parse(text, read='mysql')
+        except Exception as e:
+            statements = []
+        if len(statements) == 1 and statements[0]:
+            pretty_text = statements[0].sql(pretty=True, pad=4, dialect='mysql')
+        else:
+            pretty_text = ''
+            self.toolbar_error_message = 'Prettify failed to parse statement'
+        if len(pretty_text) > 0:
+            pretty_text = pretty_text + ';'
+        return pretty_text
+
+    def handle_unprettify_binding(self, text):
+        try:
+            statements = sqlglot.parse(text, read='mysql')
+        except Exception as e:
+            statements = []
+        if len(statements) == 1 and statements[0]:
+            unpretty_text = statements[0].sql(pretty=False, dialect='mysql')
+        else:
+            unpretty_text = ''
+            self.toolbar_error_message = 'Unprettify failed to parse statement'
+        if len(unpretty_text) > 0:
+            unpretty_text = unpretty_text + ';'
+        return unpretty_text
+
     def run_cli(self):
         iterations = 0
         sqlexecute = self.sqlexecute
@@ -613,7 +645,7 @@ class MyCli(object):
         def get_message():
             prompt = self.get_prompt(self.prompt_format)
             if self.prompt_format == self.default_prompt and len(prompt) > self.max_len_prompt:
-                prompt = self.get_prompt('\\d> ')
+                prompt = self.get_prompt(self.default_prompt_splitln)
             prompt = prompt.replace("\\x1b", "\x1b")
             return ANSI(prompt)
 
@@ -724,7 +756,7 @@ class MyCli(object):
                         except KeyboardInterrupt:
                             pass
                         if self.beep_after_seconds > 0 and t >= self.beep_after_seconds:
-                            self.echo('\a', err=True, nl=False)
+                            self.bell()
                         if special.is_timing_enabled():
                             self.echo('Time: %0.03fs' % t)
                     except KeyboardInterrupt:
@@ -865,6 +897,11 @@ class MyCli(object):
         self.log_output(s)
         click.secho(s, **kwargs)
 
+    def bell(self):
+        """Print a bell on the stderr.
+        """
+        click.secho('\a', err=True, nl=False)
+
     def get_output_margin(self, status=None):
         """Get the output margin (number of rows for the prompt, footer and
         timing message."""
@@ -938,8 +975,9 @@ class MyCli(object):
             os.environ['LESS'] = '-RXF'
 
         cnf = self.read_my_cnf_files(self.cnf_files, ['pager', 'skip-pager'])
-        if cnf['pager']:
-            special.set_pager(cnf['pager'])
+        cnf_pager = cnf['pager'] or self.config['main']['pager']
+        if cnf_pager:
+            special.set_pager(cnf_pager)
             self.explicit_pager = True
         else:
             self.explicit_pager = False
@@ -1002,7 +1040,7 @@ class MyCli(object):
         for result in results:
             title, cur, headers, status = result
             self.formatter.query = query
-            output = self.format_output(title, cur, headers)
+            output = self.format_output(title, cur, headers, special.is_expanded_output())
             for line in output:
                 click.echo(line, nl=new_line)
 
@@ -1089,6 +1127,8 @@ class MyCli(object):
 @click.option('--ssh-config-path', help='Path to ssh configuration.',
               default=os.path.expanduser('~') + '/.ssh/config')
 @click.option('--ssh-config-host', help='Host to connect to ssh server reading from ssh configuration.')
+@click.option('--ssl', 'ssl_enable', is_flag=True,
+        help='Enable SSL for connection (automatically enabled with other flags).')
 @click.option('--ssl-ca', help='CA file in PEM format.',
               type=click.Path(exists=True))
 @click.option('--ssl-capath', help='CA directory.')
@@ -1115,7 +1155,7 @@ class MyCli(object):
         help='list of DSN configured into the [alias_dsn] section of myclirc file.')
 @click.option('--list-ssh-config', 'list_ssh_config', is_flag=True,
               help='list ssh configurations in the ssh config (requires paramiko).')
-@click.option('-R', '--prompt', 'prompt',
+@click.option('-R', '--prompt', 'prompt', default=MyCli.default_prompt,
               help='Prompt format (Default: "{0}").'.format(
                   MyCli.default_prompt))
 @click.option('-l', '--logfile', type=click.File(mode='a', encoding='utf-8'),
@@ -1150,7 +1190,7 @@ class MyCli(object):
 def cli(database, user, host, port, socket, password, dbname,
         version, verbose, prompt, logfile, defaults_group_suffix,
         defaults_file, login_path, auto_vertical_output, local_infile,
-        ssl_ca, ssl_capath, ssl_cert, ssl_key, ssl_cipher,
+        ssl_enable, ssl_ca, ssl_capath, ssl_cert, ssl_key, ssl_cipher,
         tls_version, ssl_verify_server_cert, table, csv, warn, execute,
         myclirc, dsn, list_dsn, ssh_user, ssh_host, ssh_port, ssh_password,
         ssh_key_filename, list_ssh_config, ssh_config_path, ssh_config_host,
@@ -1205,6 +1245,7 @@ def cli(database, user, host, port, socket, password, dbname,
     database = dbname or database
 
     ssl = {
+            'enable': ssl_enable,
             'ca': ssl_ca and os.path.expanduser(ssl_ca),
             'cert': ssl_cert and os.path.expanduser(ssl_cert),
             'key': ssl_key and os.path.expanduser(ssl_key),
@@ -1243,7 +1284,7 @@ def cli(database, user, host, port, socket, password, dbname,
         uri = urlparse(dsn_uri)
         if not database:
             database = uri.path[1:]  # ignore the leading fwd slash
-        if not user:
+        if not user and uri.username is not None:
             user = unquote(uri.username)
         if not password and uri.password is not None:
             password = unquote(uri.password)
@@ -1296,7 +1337,12 @@ def cli(database, user, host, port, socket, password, dbname,
         try:
             if csv:
                 mycli.formatter.format_name = 'csv'
-            elif not table:
+                if execute.endswith(r'\G'):
+                    execute = execute[:-2]
+            elif table:
+                if execute.endswith(r'\G'):
+                    execute = execute[:-2]
+            else:
                 mycli.formatter.format_name = 'tsv'
 
             mycli.run_query(execute)