+++ test/test_naive_completion.py
@@ -21,7 +21,7 @@ def test_empty_string_completion(completer, complete_event):
     result = list(completer.get_completions(
         Document(text=text, cursor_position=position),
         complete_event))
-    assert result == list(map(Completion, sorted(completer.all_completions)))
+    assert result == list(map(Completion, completer.all_completions))
 
 
 def test_select_keyword_completion(completer, complete_event):
@@ -39,9 +39,7 @@ def test_function_name_completion(completer, complete_event):
     result = list(completer.get_completions(
         Document(text=text, cursor_position=position),
         complete_event))
-    assert result == list([
-        Completion(text='MASTER', start_position=-2),
-        Completion(text='MAX', start_position=-2)])
+    assert sorted(x.text for x in result) == ["MASTER", "MAX"]
 
 
 def test_column_name_completion(completer, complete_event):
@@ -50,7 +48,7 @@ def test_column_name_completion(completer, complete_event):
     result = list(completer.get_completions(
         Document(text=text, cursor_position=position),
         complete_event))
-    assert result == list(map(Completion, sorted(completer.all_completions)))
+    assert result == list(map(Completion, completer.all_completions))
 
 
 def test_special_name_completion(completer, complete_event):