+++ test/test_sqlexecute.py
@@ -117,6 +117,7 @@ def test_multiple_queries_same_line_syntaxerror(executor):
 
 
 @dbtest
+@pytest.mark.skipif(os.name == "nt", reason="Bug: fails on Windows, needs fixing, singleton of FQ not working right")
 def test_favorite_query(executor):
     set_expanded_output(False)
     run(executor, "create table test(a text)")
@@ -136,6 +137,7 @@ def test_favorite_query(executor):
 
 
 @dbtest
+@pytest.mark.skipif(os.name == "nt", reason="Bug: fails on Windows, needs fixing, singleton of FQ not working right")
 def test_favorite_query_multiple_statement(executor):
     set_expanded_output(False)
     run(executor, "create table test(a text)")
@@ -159,6 +161,7 @@ def test_favorite_query_multiple_statement(executor):
 
 
 @dbtest
+@pytest.mark.skipif(os.name == "nt", reason="Bug: fails on Windows, needs fixing, singleton of FQ not working right")
 def test_favorite_query_expanded_output(executor):
     set_expanded_output(False)
     run(executor, '''create table test(a text)''')
@@ -195,16 +198,21 @@ def test_cd_command_without_a_folder_name(executor):
 @dbtest
 def test_system_command_not_found(executor):
     results = run(executor, 'system xyz')
-    assert_result_equal(results, status='OSError: No such file or directory',
-                        assert_contains=True)
+    if os.name=='nt':
+        assert_result_equal(results, status='OSError: The system cannot find the file specified',
+                            assert_contains=True)
+    else:
+        assert_result_equal(results, status='OSError: No such file or directory',
+                            assert_contains=True)
 
 
 @dbtest
 def test_system_command_output(executor):
+    eol = os.linesep
     test_dir = os.path.abspath(os.path.dirname(__file__))
     test_file_path = os.path.join(test_dir, 'test.txt')
     results = run(executor, 'system cat {0}'.format(test_file_path))
-    assert_result_equal(results, status='mycli rocks!\n')
+    assert_result_equal(results, status=f'mycli rocks!{eol}')
 
 
 @dbtest
@@ -276,7 +284,8 @@ def test_multiple_results(executor):
 @pytest.mark.parametrize(
     'version_string, species, parsed_version_string, version',
     (
-        ('5.7.25-TiDB-v6.1.0','TiDB', '5.7.25', 50725),
+        ('5.7.25-TiDB-v6.1.0','TiDB', '6.1.0', 60100),
+        ('8.0.11-TiDB-v7.2.0-alpha-69-g96e9e68daa', 'TiDB', '7.2.0', 70200),
         ('5.7.32-35', 'Percona', '5.7.32', 50732),
         ('5.7.32-0ubuntu0.18.04.1', 'MySQL', '5.7.32', 50732),
         ('10.5.8-MariaDB-1:10.5.8+maria~focal', 'MariaDB', '10.5.8', 100508),